requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.26.0
scipy>=1.12.0
matplotlib>=3.8.0
seaborn>=0.13.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0